--- my-app/src/components/index.ts
@@ -1,3 +1,0 @@
-export * from "./footer";
-export * from "./header";
-
