--- my-app/src/components/itemCard/index.ts
@@ -1,1 +1,0 @@
-export * from './ItemCard';
