--- my-app/src/components/phoneCard/PhoneCard.types.ts
@@ -1,27 +1,0 @@
-type Phone = {
-  phone: { 
-    id: string;
-    namespaceId: string;
-    name: string;
-    capacityAvailable: string[];
-    capacity: string;
-    priceRegular: number;
-    priceDiscount: number;
-    colorsAvailable: string[];
-    color: string;
-    images: string[];
-    description: {
-      title: string;
-      text: string[];
-    }[];
-    screen: string;
-    resolution: string;
-    processor: string;
-    ram: string;
-    camera: string;
-    zoom: string;
-    cell: string[];
-  };
-};
-
-export default Phone;
